+++ src/shared/components/ui/card.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { cn } from '@/shared/lib/utils';
+import { cn } from '@/shared/lib/ui-utils';
 
 function Card({ className, ...props }: React.ComponentProps<'div'>) {
   return (
