+++ src/shared/components/ui/skeleton.tsx
@@ -1,4 +1,4 @@
-import { cn } from '@/shared/lib/utils';
+import { cn } from '@/shared/lib/ui-utils';
 
 function Skeleton({ className, ...props }: React.ComponentProps<'div'>) {
   return <div data-slot="skeleton" className={cn('animate-pulse rounded-md bg-accent', className)} {...props} />;
