+++ src/shared/lib/ui-utils.ts
